health checkers: moving more code to extensions
