+++ source/common/upstream/BUILD
@@ -208,12 +208,12 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "health_checker_base_lib",
-    srcs = ["health_checker_base_impl.cc"],
-    hdrs = ["health_checker_base_impl.h"],
+    name = "health_checker_event_logger_lib",
+    srcs = ["health_checker_event_logger.cc"],
+    hdrs = ["health_checker_event_logger.h"],
     deps = [
         "//envoy/upstream:health_checker_interface",
-        "//source/common/router:router_lib",
+        "//source/common/network:utility_lib",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/data/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/type/matcher:pkg_cc_proto",
@@ -228,13 +228,14 @@ envoy_cc_library(
         "grpc_health_proto",
     ],
     deps = [
-        ":health_checker_base_lib",
+        ":health_checker_event_logger_lib",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/data/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/type/v3:pkg_cc_proto",
         # TODO(dio): Remove dependency to server.
         "//envoy/server:health_checker_config_interface",
         "//source/common/grpc:codec_lib",
+        "//source/common/router:router_lib",
         "//source/common/http:codec_client_lib",
         "//source/common/upstream:host_utility_lib",
     ],
