+++ source/extensions/health_checkers/common/health_checker_base_impl.h
@@ -39,23 +39,6 @@ struct HealthCheckerStats {
 };
 
 /**
- * HealthCheckerHash and HealthCheckerEqualTo are used to allow the HealthCheck proto to be used as
- * a flat_hash_map key.
- */
-struct HealthCheckerHash {
-  size_t operator()(const envoy::config::core::v3::HealthCheck& health_check) const {
-    return MessageUtil::hash(health_check);
-  }
-};
-
-struct HealthCheckerEqualTo {
-  bool operator()(const envoy::config::core::v3::HealthCheck& lhs,
-                  const envoy::config::core::v3::HealthCheck& rhs) const {
-    return Protobuf::util::MessageDifferencer::Equals(lhs, rhs);
-  }
-};
-
-/**
  * Base implementation for all health checkers.
  */
 class HealthCheckerImplBase : public HealthChecker,
@@ -180,35 +163,6 @@ private:
   const std::shared_ptr<const Network::TransportSocketOptionsImpl> transport_socket_options_;
   const MetadataConstSharedPtr transport_socket_match_metadata_;
   const Common::CallbackHandlePtr member_update_cb_;
-};
-
-class HealthCheckEventLoggerImpl : public HealthCheckEventLogger {
-public:
-  HealthCheckEventLoggerImpl(AccessLog::AccessLogManager& log_manager, TimeSource& time_source,
-                             const std::string& file_name)
-      : time_source_(time_source), file_(log_manager.createAccessLog(Filesystem::FilePathAndType{
-                                       Filesystem::DestinationType::File, file_name})) {}
-
-  void logEjectUnhealthy(envoy::data::core::v3::HealthCheckerType health_checker_type,
-                         const HostDescriptionConstSharedPtr& host,
-                         envoy::data::core::v3::HealthCheckFailureType failure_type) override;
-  void logAddHealthy(envoy::data::core::v3::HealthCheckerType health_checker_type,
-                     const HostDescriptionConstSharedPtr& host, bool first_check) override;
-  void logUnhealthy(envoy::data::core::v3::HealthCheckerType health_checker_type,
-                    const HostDescriptionConstSharedPtr& host,
-                    envoy::data::core::v3::HealthCheckFailureType failure_type,
-                    bool first_check) override;
-  void logDegraded(envoy::data::core::v3::HealthCheckerType health_checker_type,
-                   const HostDescriptionConstSharedPtr& host) override;
-  void logNoLongerDegraded(envoy::data::core::v3::HealthCheckerType health_checker_type,
-                           const HostDescriptionConstSharedPtr& host) override;
-
-private:
-  void createHealthCheckEvent(
-      envoy::data::core::v3::HealthCheckerType health_checker_type, const HostDescription& host,
-      std::function<void(envoy::data::core::v3::HealthCheckEvent&)> callback) const;
-  TimeSource& time_source_;
-  AccessLog::AccessLogFileSharedPtr file_;
 };
 
 } // namespace Upstream
